Added path-based reference resolving in Grammars. (the old way of doing things is still enabled too).
You can now make crosslinks by specifying a path reference in
a grammar. For instance, in grammar we have

[Call] rule:/rules[it]

Which means find the rule named "it" (the identifier that is parsed)
in the rules field found in the root (/).

Another example, is inverses in schema:

"/" inverse:./type/fields[it]^

This means find the inverse by dereferencing the type field in
the current object (a Field), then the fields field and find "it"
in that collection.

Paths will support .. (parent) too in order to move up the spine.

NB: Not everything is updated yet.