Fixed gsub in enso.js to behave as Ruby's gsub
Gsub() in enso.js was previously aliased to replace(), but unlike replace, gsub
substitutes all occurences in the original string. This behavior was changed in
this patch. Also, note that repeated applications of replace() does not work
when the new string contains the substituted string, eg "aaa".replace("a"," a")